refactor: split compiler cli entrypoint
