+++ src/linker.rs
@@ -1,0 +1,107 @@
+use std::path::Path;
+use std::process::{self, Command};
+
+use crate::codegen::platform::{Platform, Target};
+
+pub(crate) fn assemble(target: Target, asm_path: &Path, obj_path: &Path) {
+    let mut as_cmd = Command::new(target.assembler_cmd());
+    if target.platform == Platform::MacOS {
+        as_cmd.args(["-arch", target.darwin_arch_name()]);
+    }
+    as_cmd.arg("-o").arg(obj_path).arg(asm_path);
+    run_tool("Assembler", &mut as_cmd);
+}
+
+pub(crate) fn link(
+    target: Target,
+    bin_path: &Path,
+    obj_path: &Path,
+    runtime_object_path: &Path,
+    extra_link_libs: &[String],
+    extra_link_paths: &[String],
+    extra_frameworks: &[String],
+) {
+    let mut ld_cmd = match target.platform {
+        Platform::MacOS => {
+            let sdk_path = macos_sdk_path();
+            let sdk_version = macos_sdk_version();
+            let mut cmd = Command::new("ld");
+            cmd.args(["-arch", target.darwin_arch_name(), "-e", "_main", "-o"]);
+            cmd.arg(bin_path);
+            cmd.arg(obj_path);
+            cmd.arg(runtime_object_path);
+            cmd.args(["-lSystem", "-syslibroot"]);
+            cmd.arg(&sdk_path);
+            cmd.args(["-platform_version", "macos", &sdk_version, &sdk_version]);
+            cmd
+        }
+        Platform::Linux => {
+            let mut cmd = Command::new(target.linker_cmd());
+            cmd.arg("-o").arg(bin_path).arg(obj_path).arg(runtime_object_path);
+            if extra_link_libs.is_empty() {
+                cmd.arg("-static");
+            }
+            if !extra_link_libs.is_empty() {
+                cmd.arg("-Wl,--no-as-needed");
+            }
+            cmd.args(["-lm", "-lpthread"]);
+            cmd
+        }
+    };
+    for path in extra_link_paths {
+        ld_cmd.arg(format!("-L{}", path));
+    }
+    for lib in extra_link_libs {
+        if lib != "System" {
+            ld_cmd.arg(format!("-l{}", lib));
+        }
+    }
+    if target.platform == Platform::Linux && !extra_link_libs.is_empty() {
+        ld_cmd.arg("-Wl,--as-needed");
+    }
+    if target.platform == Platform::MacOS {
+        for fw in extra_frameworks {
+            ld_cmd.args(["-framework", fw]);
+        }
+    }
+    run_tool("Linker", &mut ld_cmd);
+}
+
+fn run_tool(name: &str, cmd: &mut Command) {
+    match cmd.status() {
+        Ok(s) if s.success() => {}
+        Ok(s) => {
+            eprintln!("{} failed with exit code {}", name, s);
+            process::exit(1);
+        }
+        Err(e) => {
+            eprintln!("Failed to run {}: {}", name, e);
+            process::exit(1);
+        }
+    }
+}
+
+fn macos_sdk_path() -> String {
+    Command::new("xcrun")
+        .args(["--show-sdk-path"])
+        .output()
+        .map(|o| String::from_utf8_lossy(&o.stdout).trim().to_string())
+        .unwrap_or_default()
+}
+
+fn macos_sdk_version() -> String {
+    match Command::new("xcrun")
+        .args(["--sdk", "macosx", "--show-sdk-version"])
+        .output()
+    {
+        Ok(output) => {
+            let version = String::from_utf8_lossy(&output.stdout).trim().to_string();
+            if version.is_empty() {
+                "15.0".to_string()
+            } else {
+                version
+            }
+        }
+        Err(_) => "15.0".to_string(),
+    }
+}
